controlmaster: add timeout to run function
Pretty simple, just check if the timeout is expired whenever we check
for I/O.

Also set the timeout on the select() call to 1 second instead of 1000. I
suspect I may have mistakenly assumed this argument to be milliseconds
when I initially wrote this, so this makes more sense anyways.

A thing to note is that the timeout on select() is now effectively the
granularity with which our timeout works, so currently sub-second
timeouts are not supported -- but I don't think we desperately need
that.